updated recursion idea
doesn't crash anymore. not generating stls, creates a list of rigid
groups (unifies rigid groups and rigid joints to the same link) -
probably not correct because of None joint references not really
pointing to base, but base of component perhaps? not clear since some
joints are pointing None to None.