better error catching
improved error catching when going over joints. It will try to parse it,
if it can't it will log it. Also, put some asserts because I was sending
some floats to etree and it was not happy, so I just thought I'd leave
them there, in case in  the future some other schmuck tries to break
something else (as I did...). Now, even if it fails when going over
joints, it will still generate some URDF - that hopefully can be easily
corrected. Also tried to fix the delete function, to no avail just yet.
I probably need to read things from the table, see the item's number,
then go over the dictionary elements and then remove it. Also, to be
able to add elements in the right position (and not at the end), I
probably should do this too.