made some slight changes in the interface fixed delete button
delete button and joint;link references and renaming seem to be working
now. also i was not allowing the person to rename the package because
there was no catch for its rename command, this is also fixed.
could not fix bug that makes OK button disappear, but I put a warning on
this condition when create tree is called.